Updated encrypting algrorithm

+++ database/migrations/2015_10_07_143935_create_hits_table.php
@@ -14,9 +14,9 @@ class CreateHitsTable extends Migration
     {
        Schema::create('hits',function (Blueprint $table){
 
-           $table->increments('id');
+           $table->bigIncrements('id');
            $table->string('url_ip');
-           $table->integer('url_id')->unsigned();
+           $table->bigInteger('url_id')->unsigned();
            $table->timestamps();
 
            $table->foreign('url_id')
